Certain special-purpose value checking routines out of XMLTagHandler
