+++ src/Track.h
@@ -261,11 +261,13 @@ class AUDACITY_DLL_API Track /* not final */
 
    //! Alias for my base type
    using AttachedObjects = ::AttachedTrackObjects;
-   using ChannelType = XMLValueChecker::ChannelType;
 
-   static const auto LeftChannel = XMLValueChecker::LeftChannel;
-   static const auto RightChannel = XMLValueChecker::RightChannel;
-   static const auto MonoChannel = XMLValueChecker::MonoChannel;
+   enum ChannelType
+   {
+      LeftChannel = 0,
+      RightChannel = 1,
+      MonoChannel = 2
+   };
    
    TrackId GetId() const { return mId; }
  private:
